move redirect to cases.case.index
