+++ app/routes/cases/case.js
@@ -39,20 +39,4 @@ export default class CasesCaseRoute extends Route {
       }),
     });
   }
-
-  async redirect(model, transition) {
-    if (transition.to.name === 'cases.case.index') {
-      if (
-        model.parliamentFlow &&
-        (model.latestParliamentRetrievalActivity?.startDate >
-          model.subcases?.lastObject.created ||
-          model.latestParliamentSubmissionActivity?.startDate >
-            model.subcases?.lastObject.created)
-      ) {
-        this.router.transitionTo('cases.case.parliament-flow');
-      } else {
-        this.router.transitionTo('cases.case.subcases');
-      }
-    }
-  }
 }
